use a provided SSH key rather than generating one
Since we want the Jenkins instance to be replaceable, we want to store
the SSH key centrally (in the Vault). Having the public key stored as a
variable will also allow it to be added to the destination machines'
authorized keys easily. The downside is that the user needs to do more
setup than just allowing the key to be created on the Jenkins instance.